app/requirements.txt
fastapi>=0.115.12
uvicorn>=0.34.0
sqlalchemy>=2.0.0
pydantic>=2.0.0
pydantic-settings>=2.0.0
python-dotenv>=1.0.0
alembic>=1.11.0
pytest>=7.0.0
httpx>=0.24.0
python-multipart>=0.0.6